Defer character encoding to tablib
The previous logic ensured that all strings were bytestrings.
Unfortuantely, this breaks with e.g. the XSLX output support any time
you emit data which contains a non-ASCII character.

This patch simply ensures that everything returned is a Unicode instance
and lets tablib or other downstream consumers decide how to encode it.

Additionally, it uses tablib.compat.unicode instead of the unicode()
built-in for future compatibility with Python 3.